requirements.txt
langchain
openai
streamlit
python-dotenv
langchain-community
langchain-openai
watchdog